Master Archive
This should be seen as the scaffolding for our project - not a blueprint. If you see a file that pertains to your area of responsibility - feel free to modify it and get it working. It may be a good idea to place the database within the actual project folder in the future for ease of testing while we are developing.