Master Archive + Latest Database
Slight change. Session data garbage collection (session.clear()) added to log out actions on homepages. This clears the filter settings from one login to the next, so that they are in default settings for everyone upon initial login. This doesn't change functionality, but it makes the interface easier to navigate.